[refactor] 패키지 구조 변경 및 클래스 네이밍 수정 #7
패키지 구조를 도메인형으로 설계
- 각 도메인의 역할에 따라 구조 파악이 쉬움
- 추후에 MSA로 전환 시 분리하기 수월함
