+++ src/main/java/com/project/study/authentication_service/domain/user/oauth2/OAuth2Response.java
@@ -1,0 +1,8 @@
+package com.project.study.authentication_service.domain.user.oauth2;
+
+
+import com.project.study.member_service.domain.member.Member;
+
+public interface OAuth2Response {
+    Member toMember();
+}
