+++ src/main/java/com/project/study/member_service/domain/validation/MemberEmail.java
@@ -1,0 +1,24 @@
+package com.project.study.member_service.domain.validation;
+
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import org.hibernate.validator.constraints.Length;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Constraint(validatedBy = { })
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.FIELD, ElementType.PARAMETER})
+@NotBlank
+@Length(min = 2, max = 30)
+@Email
+public @interface MemberEmail {
+    String message() default "이메일 형식을 지켜주세요.";
+    Class<?>[] groups() default {};
+    Class<? extends Payload>[] payload() default {};
+}
