--- src/main/java/com/project/study/security/jwt/TokenType.java
@@ -1,8 +1,0 @@
-package com.project.study.security.jwt;
-
-/**
- * 토큰 타입을 구분하기 위한 ENUM
- */
-public enum TokenType {
-    ACCESS_TOKEN, REFRESH_TOKEN, OAUTH2_TOKEN
-}
