--- src/main/java/com/project/study/security/model/oauth2/OAuth2Response.java
@@ -1,7 +1,0 @@
-package com.project.study.security.model.oauth2;
-
-import com.project.study.domain.member.Member;
-
-public interface OAuth2Response {
-    Member toMember();
-}
